Renamed new-agents directory to new_agents
