+++ server/webapp/WEB-INF/rails/webpack/views/pages/new_agents/index.scss
