+++ server/webapp/WEB-INF/rails/webpack/views/pages/new_agents/spec/agent_status_widget_spec.tsx
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import {Agent} from "models/new-agent/agents";
-import {AgentsTestData} from "models/new-agent/spec/agents_test_data";
-import {AgentStatusWidget} from "views/pages/new-agents/agent_status_widget";
+import {Agent} from "models/new_agent/agents";
+import {AgentsTestData} from "models/new_agent/spec/agents_test_data";
+import {AgentStatusWidget} from "views/pages/new_agents/agent_status_widget";
 import {TestHelper} from "views/pages/spec/test_helper";
 import m from "mithril";
 import Stream from "mithril/stream";
